Fix resource merge ordering and converter crash during teardown
- Merge each resource dictionary into Application.Resources immediately
  after creation (not batch-then-merge). This ensures Styles.xaml can
  resolve {StaticResource Primary} from Colors.xaml during its
  InitializeComponent().
- Replace Application.Current.Resources lookups in RunStatusToColorConverter
  with hardcoded colors. Resources are now page-level scoped so the
  converter cannot reach them via Application.Resources, and during app
  teardown the lookup would throw KeyNotFoundException.
