+++ sample/src/DeviceTestingKitApp/MauiProgram.cs
@@ -1,4 +1,7 @@
-﻿using Microsoft.Extensions.Logging;
+﻿using DeviceTestingKitApp.Features;
+using DeviceTestingKitApp.ViewModels;
+
+using Microsoft.Extensions.Logging;
 
 namespace DeviceTestingKitApp;
 
@@ -16,5 +19,27 @@ public static class MauiProgram
 #endif
 
 		return builder.Build();
+	}
+
+	public static MauiAppBuilder AddDeviceTestingKitAppServices(this MauiAppBuilder builder)
+	{
+		builder.ConfigureFonts(fonts =>
+		{
+			fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
+			fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
+		});
+
+		// app pages
+		builder.Services.AddTransient<MainPage>();
+
+		// maui class library
+		builder.Services.AddTransient<ISemanticScreenReader>(_ => SemanticScreenReader.Default);
+		builder.Services.AddTransient<ISemanticAnnouncer, MauiSemanticAnnouncer>();
+
+		// plain class library
+		builder.Services.AddTransient<MainViewModel>();
+		builder.Services.AddTransient<CounterViewModel>();
+
+		return builder;
 	}
 }
